// projects/angular-sdk/src/apis/hopper-cloud-airline/v1/model/checkCfarContractExerciceVerificationCodeResponse.ts
/**
 * Airline API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface CheckCfarContractExerciceVerificationCodeResponse { 
    compliant: boolean;
    /**
     * Unique identifier for a contract
     */
    contractId: string;
    /**
     * Unique identifier for a CFAR exercise
     */
    exerciseId: string;
}