// projects/angular-sdk/src/apis/hopper-cloud-airline/v1/model/checkCfarContractExerciseVerificationCodeRequest.ts
/**
 * Airline API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * A request associated with validation code checks before processing a CFAR contract exercise
 */
export interface CheckCfarContractExerciseVerificationCodeRequest { 
    verificationCode: string;
}